Update calorie calculations

+++ TrainingTrackerAPI/Models/Activity.cs
@@ -13,7 +13,7 @@ namespace TrainingTrackerAPI.Models
         public double Distance { get; set; }
         public int TotalTimeInSeconds { get; set; }
         public TimeOnly TimeInput { get; set; }
-        public int? CaloriesBurned { get; set; }
+        public double? CaloriesBurned { get; set; }
         public ApplicationUser? User { get; set; }
 
         public string? UserId { get; set; }
